# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carpool Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/users/user-info": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get auth user",
                "description": "get auth user ",
                "operationId": "getUserInfo",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/send-code": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "send code again",
                "description": "send code again",
                "operationId": "sendCode",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "mobile",
                                        "example": "09156326123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/check-code": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "check user code",
                "description": "check user code",
                "operationId": "checkUserCode",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "mobile",
                                        "example": "09156326123"
                                    },
                                    "code": {
                                        "description": "code",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/complete-profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "complete user profile",
                "description": "complete user profile",
                "operationId": "createOrUpdateUserData",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "example": "mehdi"
                                    },
                                    "family": {
                                        "description": "family",
                                        "example": "firoozi"
                                    },
                                    "birth_day": {
                                        "description": "birth_day",
                                        "example": "1995-12-12"
                                    },
                                    "gender": {
                                        "description": "gender",
                                        "example": "1"
                                    },
                                    "national_code": {
                                        "description": "national_code",
                                        "example": "2123134145"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/update-bio": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "update user bio",
                "description": "update user bio",
                "operationId": "updateUserBio",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "description": "text",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/upload-file": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "upload user file",
                "description": "upload user file",
                "operationId": "uploadUserFile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image",
                                        "type": "file",
                                        "example": "test"
                                    },
                                    "type": {
                                        "description": "type",
                                        "example": "profile,drive_license_back,drive_license_front,selfie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/update-preference/{preferenceId}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "update user preference",
                "description": "update user preference",
                "operationId": "updateUserPreference",
                "parameters": [
                    {
                        "name": "preferenceId",
                        "in": "path",
                        "description": "preference id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "preference_option_id": {
                                        "description": "preference_option_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "",
                "name": "Authorization",
                "in": "header",
                "description": "Bearer token for Sanctum authentication"
            }
        }
    ]
}